One attempt at LiDAR
This is somehow the only attempt at getting the LiDAR sensor to work that I had on my computer. I guess all of the other attempts are in VSCode on one of the Raspberry Pis. Just look for any .py files and see if they look similar to this. None of them work fully with the sensor we bought. Support is limited for that waterproof one. It will be easier to work on the exposed $100 sensor and you can work up from there.